requirements.txt
Flask==3.0.0
gunicorn==21.2.0
pytest==7.4.3
Flask-Talisman==1.1.0
google-cloud-logging==3.8.0
